Glob test moved to util.ts tests
